pkg/cluster: split gateway options into separate Go files
Instead of one mega function for all three gateway options,
put each on in its own file.
