+++ go-controller/pkg/cluster/gateway_init.go
@@ -5,32 +5,12 @@ package cluster
 import (
 	"fmt"
 	"net"
-	"regexp"
-	"strings"
 	"syscall"
-
-	"github.com/openvswitch/ovn-kubernetes/go-controller/pkg/config"
-	"github.com/openvswitch/ovn-kubernetes/go-controller/pkg/factory"
 
 	"github.com/coreos/go-iptables/iptables"
 	"github.com/openvswitch/ovn-kubernetes/go-controller/pkg/util"
-	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
-	kapi "k8s.io/api/core/v1"
-	"k8s.io/client-go/tools/cache"
 )
-
-const (
-	localnetGatewayIP            = "169.254.33.2/24"
-	localnetGatewayNextHop       = "169.254.33.1"
-	localnetGatewayNextHopSubnet = "169.254.33.1/24"
-)
-
-type iptRule struct {
-	table string
-	chain string
-	args  []string
-}
 
 // getIPv4Address returns the ipv4 address for the network interface 'iface'.
 func getIPv4Address(iface string) (string, error) {
@@ -54,68 +34,6 @@ func getIPv4Address(iface string) (string, error) {
 		}
 	}
 	return ipAddress, nil
-}
-
-func ensureChain(ipt util.IPTablesHelper, table, chain string) error {
-	chains, err := ipt.ListChains(table)
-	if err != nil {
-		return fmt.Errorf("failed to list iptables chains: %v", err)
-	}
-	for _, ch := range chains {
-		if ch == chain {
-			return nil
-		}
-	}
-
-	return ipt.NewChain(table, chain)
-}
-
-func generateGatewayNATRules(ifname string, ip string) []iptRule {
-	// Allow packets to/from the gateway interface in case defaults deny
-	rules := make([]iptRule, 0)
-	rules = append(rules, iptRule{
-		table: "filter",
-		chain: "FORWARD",
-		args:  []string{"-i", ifname, "-j", "ACCEPT"},
-	})
-	rules = append(rules, iptRule{
-		table: "filter",
-		chain: "FORWARD",
-		args: []string{"-o", ifname, "-m", "conntrack", "--ctstate",
-			"RELATED,ESTABLISHED", "-j", "ACCEPT"},
-	})
-	rules = append(rules, iptRule{
-		table: "filter",
-		chain: "INPUT",
-		args:  []string{"-i", ifname, "-m", "comment", "--comment", "from OVN to localhost", "-j", "ACCEPT"},
-	})
-
-	// NAT for the interface
-	rules = append(rules, iptRule{
-		table: "nat",
-		chain: "POSTROUTING",
-		args:  []string{"-s", ip, "-j", "MASQUERADE"},
-	})
-	return rules
-}
-
-func localnetGatewayNAT(ipt util.IPTablesHelper, ifname, ip string) error {
-	rules := generateGatewayNATRules(ifname, ip)
-	for _, r := range rules {
-		if err := ensureChain(ipt, r.table, r.chain); err != nil {
-			return fmt.Errorf("failed to ensure %s/%s: %v", r.table, r.chain, err)
-		}
-		exists, err := ipt.Exists(r.table, r.chain, r.args...)
-		if !exists && err == nil {
-			err = ipt.Insert(r.table, r.chain, 1, r.args...)
-		}
-		if err != nil {
-			return fmt.Errorf("failed to add iptables %s/%s rule %q: %v",
-				r.table, r.chain, strings.Join(r.args, " "), err)
-		}
-	}
-
-	return nil
 }
 
 // getDefaultGatewayInterfaceDetails returns the interface name on
@@ -143,239 +61,6 @@ func getDefaultGatewayInterfaceDetails() (string, string, error) {
 	return "", "", fmt.Errorf("Failed to get default gateway interface")
 }
 
-func addDefaultConntrackRules(gwBridge, gwIntf string) error {
-	patchPort := "k8s-patch-" + gwBridge + "-br-int"
-	// Get ofport of pathPort
-	ofportPatch, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-		"interface", patchPort, "ofport")
-	if err != nil {
-		return fmt.Errorf("Failed to get ofport of %s, stderr: %q, error: %v",
-			patchPort, stderr, err)
-	}
-
-	// Get ofport of physical interface
-	ofportPhys, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-		"interface", gwIntf, "ofport")
-	if err != nil {
-		return fmt.Errorf("Failed to get ofport of %s, stderr: %q, error: %v",
-			gwIntf, stderr, err)
-	}
-
-	// table 0, packets coming from pods headed externally. Commit connections
-	// so that reverse direction goes back to the pods.
-	_, stderr, err = util.RunOVSOfctl("add-flow", gwBridge,
-		fmt.Sprintf("priority=100, in_port=%s, ip, "+
-			"actions=ct(commit, zone=%d), output:%s",
-			ofportPatch, config.Default.ConntrackZone, ofportPhys))
-	if err != nil {
-		return fmt.Errorf("Failed to add openflow flow to %s, stderr: %q, "+
-			"error: %v", gwBridge, stderr, err)
-	}
-
-	// table 0, packets coming from external. Send it through conntrack and
-	// resubmit to table 1 to know the state of the connection.
-	_, stderr, err = util.RunOVSOfctl("add-flow", gwBridge,
-		fmt.Sprintf("priority=50, in_port=%s, ip, "+
-			"actions=ct(zone=%d, table=1)", ofportPhys, config.Default.ConntrackZone))
-	if err != nil {
-		return fmt.Errorf("Failed to add openflow flow to %s, stderr: %q, "+
-			"error: %v", gwBridge, stderr, err)
-	}
-
-	// table 1, established and related connections go to pod
-	_, stderr, err = util.RunOVSOfctl("add-flow", gwBridge,
-		fmt.Sprintf("priority=100, table=1, ct_state=+trk+est, "+
-			"actions=output:%s", ofportPatch))
-	if err != nil {
-		return fmt.Errorf("Failed to add openflow flow to %s, stderr: %q, "+
-			"error: %v", gwBridge, stderr, err)
-	}
-	_, stderr, err = util.RunOVSOfctl("add-flow", gwBridge,
-		fmt.Sprintf("priority=100, table=1, ct_state=+trk+rel, "+
-			"actions=output:%s", ofportPatch))
-	if err != nil {
-		return fmt.Errorf("Failed to add openflow flow to %s, stderr: %q, "+
-			"error: %v", gwBridge, stderr, err)
-	}
-
-	// table 1, all other connections go to the bridge interface.
-	_, stderr, err = util.RunOVSOfctl("add-flow", gwBridge,
-		"priority=0, table=1, actions=output:LOCAL")
-	if err != nil {
-		return fmt.Errorf("Failed to add openflow flow to %s, stderr: %q, "+
-			"error: %v", gwBridge, stderr, err)
-	}
-	return nil
-}
-
-func addService(service *kapi.Service, inport, outport, gwBridge string) {
-	if service.Spec.Type != kapi.ServiceTypeNodePort {
-		return
-	}
-
-	for _, svcPort := range service.Spec.Ports {
-		if svcPort.Protocol != kapi.ProtocolTCP &&
-			svcPort.Protocol != kapi.ProtocolUDP {
-			continue
-		}
-		protocol := strings.ToLower(string(svcPort.Protocol))
-
-		_, stderr, err := util.RunOVSOfctl("add-flow", gwBridge,
-			fmt.Sprintf("priority=100, in_port=%s, %s, tp_dst=%d, actions=%s",
-				inport, protocol, svcPort.NodePort, outport))
-		if err != nil {
-			logrus.Errorf("Failed to add openflow flow on %s for nodePort "+
-				"%d, stderr: %q, error: %v", gwBridge,
-				svcPort.NodePort, stderr, err)
-		}
-	}
-}
-
-func deleteService(service *kapi.Service, inport, gwBridge string) {
-	if service.Spec.Type != kapi.ServiceTypeNodePort {
-		return
-	}
-
-	for _, svcPort := range service.Spec.Ports {
-		if svcPort.Protocol != kapi.ProtocolTCP &&
-			svcPort.Protocol != kapi.ProtocolUDP {
-			continue
-		}
-
-		protocol := strings.ToLower(string(svcPort.Protocol))
-
-		_, stderr, err := util.RunOVSOfctl("del-flows", gwBridge,
-			fmt.Sprintf("in_port=%s, %s, tp_dst=%d",
-				inport, protocol, svcPort.NodePort))
-		if err != nil {
-			logrus.Errorf("Failed to delete openflow flow on %s for nodePort "+
-				"%d, stderr: %q, error: %v", gwBridge,
-				svcPort.NodePort, stderr, err)
-		}
-	}
-}
-
-func syncServices(services []interface{}, gwBridge, gwIntf string) {
-	// Get ofport of physical interface
-	inport, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-		"interface", gwIntf, "ofport")
-	if err != nil {
-		logrus.Errorf("Failed to get ofport of %s, stderr: %q, error: %v",
-			gwIntf, stderr, err)
-		return
-	}
-
-	nodePorts := make(map[string]bool)
-	for _, serviceInterface := range services {
-		service, ok := serviceInterface.(*kapi.Service)
-		if !ok {
-			logrus.Errorf("Spurious object in syncServices: %v",
-				serviceInterface)
-			continue
-		}
-
-		if service.Spec.Type != kapi.ServiceTypeNodePort ||
-			len(service.Spec.Ports) == 0 {
-			continue
-		}
-
-		for _, svcPort := range service.Spec.Ports {
-			port := svcPort.NodePort
-			if port == 0 {
-				continue
-			}
-
-			prot := svcPort.Protocol
-			if prot != kapi.ProtocolTCP && prot != kapi.ProtocolUDP {
-				continue
-			}
-			protocol := strings.ToLower(string(prot))
-			nodePortKey := fmt.Sprintf("%s_%d", protocol, port)
-			nodePorts[nodePortKey] = true
-		}
-	}
-
-	stdout, stderr, err := util.RunOVSOfctl("dump-flows",
-		gwBridge)
-	if err != nil {
-		logrus.Errorf("dump-flows failed: %q (%v)", stderr, err)
-		return
-	}
-	flows := strings.Split(stdout, "\n")
-
-	re, err := regexp.Compile(`tp_dst=(.*?)[, ]`)
-	if err != nil {
-		logrus.Errorf("regexp compile failed: %v", err)
-		return
-	}
-
-	for _, flow := range flows {
-		group := re.FindStringSubmatch(flow)
-		if group == nil {
-			continue
-		}
-
-		var key string
-		if strings.Contains(flow, "tcp") {
-			key = fmt.Sprintf("tcp_%s", group[1])
-		} else if strings.Contains(flow, "udp") {
-			key = fmt.Sprintf("udp_%s", group[1])
-		} else {
-			continue
-		}
-
-		if _, ok := nodePorts[key]; !ok {
-			pair := strings.Split(key, "_")
-			protocol, port := pair[0], pair[1]
-
-			stdout, _, err := util.RunOVSOfctl(
-				"del-flows", gwBridge,
-				fmt.Sprintf("in_port=%s, %s, tp_dst=%s",
-					inport, protocol, port))
-			if err != nil {
-				logrus.Errorf("del-flows of %s failed: %q",
-					gwBridge, stdout)
-			}
-		}
-	}
-}
-
-func nodePortWatcher(gwBridge, gwIntf string, wf *factory.WatchFactory) error {
-	patchPort := "k8s-patch-" + gwBridge + "-br-int"
-	// Get ofport of patchPort
-	ofportPatch, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-		"interface", patchPort, "ofport")
-	if err != nil {
-		return fmt.Errorf("Failed to get ofport of %s, stderr: %q, error: %v",
-			patchPort, stderr, err)
-	}
-
-	// Get ofport of physical interface
-	ofportPhys, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-		"interface", gwIntf, "ofport")
-	if err != nil {
-		return fmt.Errorf("Failed to get ofport of %s, stderr: %q, error: %v",
-			gwIntf, stderr, err)
-	}
-
-	_, err = wf.AddServiceHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			service := obj.(*kapi.Service)
-			addService(service, ofportPhys, ofportPatch, gwBridge)
-		},
-		UpdateFunc: func(old, new interface{}) {
-		},
-		DeleteFunc: func(obj interface{}) {
-			service := obj.(*kapi.Service)
-			deleteService(service, ofportPhys, gwBridge)
-		},
-	}, func(services []interface{}) {
-		syncServices(services, gwBridge, gwIntf)
-	})
-
-	return err
-}
-
 func (cluster *OvnClusterController) initGateway(
 	nodeName string, clusterIPSubnet []string, subnet string) error {
 	ipt, err := iptables.NewWithProtocol(iptables.ProtocolIPv4)
@@ -389,65 +74,7 @@ func (cluster *OvnClusterController) initGatewayInternal(
 	nodeName string, clusterIPSubnet []string, subnet string,
 	ipt util.IPTablesHelper) error {
 	if cluster.LocalnetGateway {
-		// Create a localnet OVS bridge.
-		localnetBridgeName := "br-localnet"
-		_, stderr, err := util.RunOVSVsctl("--may-exist", "add-br",
-			localnetBridgeName)
-		if err != nil {
-			return fmt.Errorf("Failed to create localnet bridge %s"+
-				", stderr:%s (%v)", localnetBridgeName, stderr, err)
-		}
-
-		_, _, err = util.RunIP("link", "set", localnetBridgeName, "up")
-		if err != nil {
-			return fmt.Errorf("failed to up %s (%v)", localnetBridgeName, err)
-		}
-
-		// Create a localnet bridge nexthop
-		localnetBridgeNextHop := "br-nexthop"
-		_, stderr, err = util.RunOVSVsctl("--may-exist", "add-port",
-			localnetBridgeName, localnetBridgeNextHop, "--", "set",
-			"interface", localnetBridgeNextHop, "type=internal")
-		if err != nil {
-			return fmt.Errorf("Failed to create localnet bridge next hop %s"+
-				", stderr:%s (%v)", localnetBridgeNextHop, stderr, err)
-		}
-		_, _, err = util.RunIP("link", "set", localnetBridgeNextHop, "up")
-		if err != nil {
-			return fmt.Errorf("failed to up %s (%v)", localnetBridgeNextHop, err)
-		}
-
-		// Flush IPv4 address of localnetBridgeNextHop.
-		_, _, err = util.RunIP("addr", "flush", "dev", localnetBridgeNextHop)
-		if err != nil {
-			return fmt.Errorf("failed to flush ip address of %s (%v)",
-				localnetBridgeNextHop, err)
-		}
-
-		// Set localnetBridgeNextHop with an IP address.
-		_, _, err = util.RunIP("addr", "add",
-			localnetGatewayNextHopSubnet,
-			"dev", localnetBridgeNextHop)
-		if err != nil {
-			return fmt.Errorf("failed to assign ip address to %s (%v)",
-				localnetBridgeNextHop, err)
-		}
-
-		err = util.GatewayInit(clusterIPSubnet, nodeName,
-			localnetGatewayIP, "",
-			localnetBridgeName, localnetGatewayNextHop, subnet,
-			false)
-		if err != nil {
-			return fmt.Errorf("failed to GatewayInit for localnet (%v)", err)
-		}
-
-		err = localnetGatewayNAT(ipt, localnetBridgeNextHop, localnetGatewayIP)
-		if err != nil {
-			return fmt.Errorf("Failed to add NAT rules for localnet gateway (%v)",
-				err)
-		}
-
-		return nil
+		return initLocalnetGateway(nodeName, clusterIPSubnet, subnet, ipt)
 	}
 
 	if cluster.GatewayNextHop == "" || cluster.GatewayIntf == "" {
@@ -466,97 +93,19 @@ func (cluster *OvnClusterController) initGatewayInternal(
 		}
 	}
 
-	var ipAddress string
-	var err error
-	if !cluster.GatewaySpareIntf {
-		// Check to see whether the interface is OVS bridge.
-		_, _, err = util.RunOVSVsctl("--", "br-exists", cluster.GatewayIntf)
-		if err != nil {
-			// This is not a OVS bridge. We need to create a OVS bridge
-			// and add cluster.GatewayIntf as a port of that bridge.
-			bridgeName, err := util.NicToBridge(cluster.GatewayIntf)
-			if err != nil {
-				return fmt.Errorf("failed to convert %s to OVS bridge: %v",
-					cluster.GatewayIntf, err)
-			}
-			cluster.GatewayBridge = bridgeName
-		} else {
-			// The given (or autodetected) interface is an OVS bridge;
-			// detect if we previously ran NIC/bridge setup
-			if !strings.HasPrefix(cluster.GatewayIntf, "br") {
-				return fmt.Errorf("gateway interface %s is an OVS bridge not "+
-					"a physical device", cluster.GatewayIntf)
-			}
-
-			// Is intfName a port of cluster.GatewayIntf?
-			intfName := util.GetNicName(cluster.GatewayIntf)
-			_, stderr, err := util.RunOVSVsctl("--if-exists", "get",
-				"interface", intfName, "ofport")
-			if err != nil {
-				return fmt.Errorf("failed to get ofport of %s, stderr: %q, error: %v",
-					intfName, stderr, err)
-			}
-			cluster.GatewayBridge = cluster.GatewayIntf
-			cluster.GatewayIntf = intfName
-		}
-
-		// Now, we get IP address from OVS bridge. If IP does not exist,
-		// error out.
-		ipAddress, err = getIPv4Address(cluster.GatewayBridge)
-		if err != nil {
-			return fmt.Errorf("Failed to get interface details for %s (%v)",
-				cluster.GatewayBridge, err)
-		}
-		if ipAddress == "" {
-			return fmt.Errorf("%s does not have a ipv4 address",
-				cluster.GatewayBridge)
-		}
-		err = util.GatewayInit(clusterIPSubnet, nodeName, ipAddress, "",
-			cluster.GatewayBridge, cluster.GatewayNextHop, subnet,
+	if cluster.GatewaySpareIntf {
+		return initSpareGateway(nodeName, clusterIPSubnet, subnet,
+			cluster.GatewayNextHop, cluster.GatewayIntf,
 			cluster.NodePortEnable)
-		if err != nil {
-			logrus.Errorf("failed to GatewayInit (%v)", err)
-			return err
-		}
-
-	} else {
-		// Now, we get IP address from physical interface. If IP does not
-		// exists error out.
-		ipAddress, err = getIPv4Address(cluster.GatewayIntf)
-		if err != nil {
-			return fmt.Errorf("Failed to get interface details for %s (%v)",
-				cluster.GatewayIntf, err)
-		}
-		if ipAddress == "" {
-			return fmt.Errorf("%s does not have a ipv4 address",
-				cluster.GatewayIntf)
-		}
-		err = util.GatewayInit(clusterIPSubnet, nodeName, ipAddress,
-			cluster.GatewayIntf, "", cluster.GatewayNextHop, subnet,
-			cluster.NodePortEnable)
-		if err != nil {
-			logrus.Errorf("failed to GatewayInit (%v)", err)
-			return err
-		}
 	}
 
-	if !cluster.GatewaySpareIntf {
-		// Program cluster.GatewayIntf to let non-pod traffic to go to host
-		// stack
-		err = addDefaultConntrackRules(cluster.GatewayBridge, cluster.GatewayIntf)
-		if err != nil {
-			return err
-		}
-
-		if cluster.NodePortEnable {
-			// Program cluster.GatewayIntf to let nodePort traffic to go to pods.
-			err = nodePortWatcher(cluster.GatewayBridge,
-				cluster.GatewayIntf, cluster.watchFactory)
-			if err != nil {
-				return err
-			}
-		}
+	bridge, gwIntf, err := initSharedGateway(nodeName, clusterIPSubnet, subnet,
+		cluster.GatewayNextHop, cluster.GatewayIntf, cluster.NodePortEnable,
+		cluster.watchFactory)
+	if err != nil {
+		return err
 	}
-
+	cluster.GatewayBridge = bridge
+	cluster.GatewayIntf = gwIntf
 	return nil
 }
